ML dependency update + coreml support (#175)
- update to ort version rc10 + fix some breaking changes
- conditional dependencies declaration for the execution provider
features of ort, to add the coreml feature on macos and cuda on linux.
Previously we could not take advantage of the coreml API on mac as
adding it as a feature to Cargo.toml caused the build to fail on non-mac
systems
- conditionally register the execution provider based on the target os
(same logic as for the dependency)
- added instrumentation to all ML worker methods for which monitoring
performance at our stage of the development can prove useful (which for
now, is most!).
- we wrap some of the code in the main loop of the `run_alert_worker`
function into their own sub-functions with tracing, so we can have better timing information.
- update the Cargo.lock overall, doesn't hurt to keep things fresh!