Customizable Kafka URLs (#203)
- adds a `kafka` section to the config, with 2 keys: `producer` and
`consumer`.
- `producer` is a simple string value, as it refers to the url where we
want to produce the output of BOOM (used by the filter workers), which
for now is just one server.
- `consumer` has one key per survey (e.g. ZTF), and their values are the
kafka servers to consume alerts from (used by the kafka consumer bin)
- adds a `KafkaConfig` struct and methods to populate it from the config
in `src::conf`, which are used in the kafka consumer and filter workers
to read these keys from the config.
- the `kafka_producer` binary (which we use to create local kafka topics
with alerts, only used for development as we do not point to real kafka
instances in that case) can now take an optional `server_url`
parameters, which is the url of the kafka server to produce too. Given
that this is a "dev-only" tool, it feels inappropriate to have this be
configurable from the config file too.