API docs improvements (#210)
In this PR, we:
- add JWT authentication information to the docs
- add tags to openapi description to have proper menus in the docs (tags
are used to group endpoints),
- add description (as comments) to some endpoints which are rendered on
the docs
- add a dedicated `docs.rs` where the documentation is instantiated.
- add a `routes/info.rs` where the ping and db info endpoints now live,
instead of being directly in main (required to be documented in
`docs.rs`)

Note: While the JWT auth is documented, the docs ask the user for their
token. Ideally, the docs should ask for username/password and somehow
know which endpoint to hit to get a JWT it can use thereafter when
testing endpoints from the web page. I couldn't find that option in
utoipa so I opened a ticket on its GitHub repo. That should be added in
a future PR